seperated Revit wall subclasses into own files
