+++ Objects/Objects/BuiltElements/Revit/RevitWall.cs
@@ -119,125 +119,3 @@ public class RevitWall : Wall
   public Base? parameters { get; set; }
   public string elementId { get; set; }
 }
-
-public class RevitFaceWall : Wall
-{
-  public RevitFaceWall() { }
-
-  [SchemaInfo("RevitWall by face", "Creates a Revit wall from a surface.", "Revit", "Architecture")]
-  public RevitFaceWall(
-    string family,
-    string type,
-    [SchemaParamInfo("Surface or single face Brep to use"), SchemaMainParam] Brep surface,
-    Level level,
-    LocationLine locationLine = LocationLine.Interior,
-    [SchemaParamInfo("Set in here any nested elements that this wall might have.")] List<Base>? elements = null,
-    List<Parameter>? parameters = null
-  )
-  {
-    if (surface.Surfaces.Count == 0)
-    {
-      throw new Exception("Cannot create a RevitWall with an empty BREP");
-    }
-
-    if (surface.Surfaces.Count > 1)
-    {
-      throw new Exception(
-        "The provided brep has more than 1 surface. Please deconstruct/explode it to create multiple instances"
-      );
-    }
-
-    this.family = family;
-    this.type = type;
-    brep = surface;
-    this.locationLine = locationLine;
-    this.level = level;
-    this.elements = elements;
-    this.parameters = parameters?.ToBase();
-  }
-
-  public string family { get; set; }
-  public string type { get; set; }
-
-  public Brep brep { get; set; }
-
-  public new Level? level
-  {
-    get => base.level;
-    set => base.level = value;
-  }
-
-  public LocationLine locationLine { get; set; }
-  public Base? parameters { get; set; }
-  public string elementId { get; set; }
-}
-
-public class RevitProfileWall : Wall
-{
-  public RevitProfileWall() { }
-
-  [SchemaInfo("RevitWall by profile", "Creates a Revit wall from a profile.", "Revit", "Architecture")]
-  public RevitProfileWall(
-    string family,
-    string type,
-    [SchemaParamInfo("Profile to use"), SchemaMainParam] Polycurve profile,
-    Level level,
-    LocationLine locationLine = LocationLine.Interior,
-    bool structural = false,
-    [SchemaParamInfo("Set in here any nested elements that this wall might have.")] List<Base>? elements = null,
-    List<Parameter>? parameters = null
-  )
-  {
-    this.family = family;
-    this.type = type;
-    this.profile = profile;
-    this.locationLine = locationLine;
-    this.structural = structural;
-    this.level = level;
-    this.elements = elements;
-    this.parameters = parameters?.ToBase();
-  }
-
-  public string family { get; set; }
-  public string type { get; set; }
-  public Polycurve profile { get; set; }
-
-  public new Level? level
-  {
-    get => base.level;
-    set => base.level = value;
-  }
-
-  public LocationLine locationLine { get; set; }
-  public bool structural { get; set; }
-  public Base? parameters { get; set; }
-  public string elementId { get; set; }
-}
-
-// [SchemaDescription("Not supported yet.")]
-// [SchemaIgnore]
-// public class RevitCurtainWall : Wall
-// {
-//   // TODO
-//   // What props do/can curtain walls have? - grid, mullions, etc.
-//
-//   [SchemaOptional]
-//   public bool flipped { get; set; }
-//
-//   [SchemaOptional]
-//   public Base parameters { get; set; }
-//
-//   [SchemaIgnore]
-//   public string elementId { get; set; }
-// }
-//
-// [SchemaDescription("Not supported yet.")]
-// [SchemaIgnore]
-// public class RevitWallByPoint : Base
-// {
-//   [SchemaOptional]
-//   public Base parameters { get; set; }
-//
-//   [SchemaIgnore]
-//   public string elementId { get; set; }
-// }
